+++ personservice/src/main/java/com/sap/demo/applicationconnector/configuration/PairingServiceConfiguration.java
@@ -1,4 +1,4 @@
-package com.sap.demo.applicationconnector.pairing;
+package com.sap.demo.applicationconnector.configuration;
 
 import org.apache.http.client.HttpClient;
 import org.apache.http.impl.client.HttpClients;
